+++ fastly/block_fastly_service_logging_https_test.go
@@ -31,6 +31,7 @@ func TestAccFastlyServiceVCL_httpslogging_basic(t *testing.T) {
 		MessageType:       gofastly.ToPointer("blank"),
 		Method:            gofastly.ToPointer("PUT"),
 		Name:              gofastly.ToPointer("httpslogger"),
+		Period:            gofastly.ToPointer(10),
 		RequestMaxBytes:   gofastly.ToPointer(0),
 		RequestMaxEntries: gofastly.ToPointer(0),
 		ResponseCondition: gofastly.ToPointer(""),
@@ -52,6 +53,7 @@ func TestAccFastlyServiceVCL_httpslogging_basic(t *testing.T) {
 		MessageType:       gofastly.ToPointer("blank"),
 		Method:            gofastly.ToPointer("POST"),
 		Name:              gofastly.ToPointer("httpslogger"),
+		Period:            gofastly.ToPointer(30),
 		RequestMaxBytes:   gofastly.ToPointer(0),
 		RequestMaxEntries: gofastly.ToPointer(0),
 		ResponseCondition: gofastly.ToPointer(""),
@@ -72,6 +74,7 @@ func TestAccFastlyServiceVCL_httpslogging_basic(t *testing.T) {
 		MessageType:       gofastly.ToPointer("blank"),
 		Method:            gofastly.ToPointer("POST"),
 		Name:              gofastly.ToPointer("httpslogger2"),
+		Period:            gofastly.ToPointer(60),
 		RequestMaxBytes:   gofastly.ToPointer(1000),
 		RequestMaxEntries: gofastly.ToPointer(0),
 		ResponseCondition: gofastly.ToPointer(""),
@@ -92,6 +95,7 @@ func TestAccFastlyServiceVCL_httpslogging_basic(t *testing.T) {
 		MessageType:       gofastly.ToPointer("blank"),
 		Method:            gofastly.ToPointer("PUT"),
 		Name:              gofastly.ToPointer("httpslogger3"),
+		Period:            gofastly.ToPointer(120),
 		RequestMaxBytes:   gofastly.ToPointer(0),
 		RequestMaxEntries: gofastly.ToPointer(0),
 		ResponseCondition: gofastly.ToPointer(""),
@@ -156,6 +160,7 @@ func TestAccFastlyServiceVCL_httpslogging_basic_compute(t *testing.T) {
 		MessageType:       gofastly.ToPointer("blank"),
 		Method:            gofastly.ToPointer("PUT"),
 		Name:              gofastly.ToPointer("httpslogger"),
+		Period:            gofastly.ToPointer(300),
 		RequestMaxBytes:   gofastly.ToPointer(0),
 		RequestMaxEntries: gofastly.ToPointer(0),
 		ServiceVersion:    gofastly.ToPointer(1),
@@ -281,6 +286,7 @@ resource "fastly_service_vcl" "foo" {
 		name               = "httpslogger3"
 		method             = "PUT"
 		format             = %q
+		period             = 120
 		url                = "https://example.com/logs/3"
 	}
 
@@ -306,6 +312,7 @@ resource "fastly_service_vcl" "foo" {
 	logging_https {
 		name               = "httpslogger"
 		method             = "PUT"
+		period			   = 10
 		url                = "https://example.com/logs/1"
 		compression_codec = "zstd"
     processing_region = "us"
@@ -339,7 +346,8 @@ resource "fastly_service_compute" "foo" {
 		method             = "PUT"
 		url                = "https://example.com/logs/1"
 		compression_codec = "zstd"
-    processing_region = "us"
+		period  		  = 300
+        processing_region = "us"
 	}
 
   package {
@@ -370,6 +378,7 @@ resource "fastly_service_vcl" "foo" {
 		name               = "httpslogger"
 		format             = %q
 		method             = "POST"
+		period			   = 30
 		url                = "https://example.com/logs/1"
 		compression_codec = "snappy"
 	}
@@ -379,6 +388,7 @@ resource "fastly_service_vcl" "foo" {
 		format             = %q
 		method             = "POST"
 		url                = "https://example.com/logs/2"
+		period			   = 60
 		request_max_bytes  = 1000
 		gzip_level				 = 5
 	}
@@ -405,6 +415,7 @@ func TestResourceFastlyFlattenHTTPS(t *testing.T) {
 					GzipLevel:         gofastly.ToPointer(0),
 					Format:            gofastly.ToPointer(LoggingHTTPSDefaultFormat),
 					FormatVersion:     gofastly.ToPointer(2),
+					Period:            gofastly.ToPointer(5),
 					ProcessingRegion:  gofastly.ToPointer("eu"),
 				},
 			},
@@ -420,6 +431,7 @@ func TestResourceFastlyFlattenHTTPS(t *testing.T) {
 					"format":              LoggingHTTPSDefaultFormat,
 					"format_version":      2,
 					"gzip_level":          0,
+					"period":              5,
 					"processing_region":   "eu",
 				},
 			},